Use ignoreerrors to ignore unknown sysctl keys
Doc says "ignoreerrors: Use this option to ignore errors about unknown keys.", but this is transmitted as "-e" option to sysctl only when doing a reload, not when setting a key.

This commit adds -e to the command line also when setting a key. This solves my problem. (In my case it's an ipv6 key when ipv6 is disabled)

However this does not detect that the key is written into a black hole, and thus returns "changed" even when nothing actually happened. Avoiding this is a lot of work, either trapping the error or checking that the action produced a change. It might be better to simply amend the docs to say that -e is not taken into account when setting a specific key unless using reload:yes